+++ src/Modules/Settings/Settings.js
@@ -15,6 +15,8 @@ import i18n from '@/helper/i18n'
 import { subscribeForPushNotifications, unsubscribeFromPushNotifications } from '@/pushNotifications'
 import { confirmDeleteUser } from '../Foodsaver/Foodsaver'
 import { vueApply, vueRegister } from '@/vue'
+import { setSleepStatus } from '@/api/user'
+import $ from 'jquery'
 import Calendar from './components/Calendar'
 import ProfilePicture from './components/ProfilePicture'
 import NameInput from './components/NameInput'
@@ -36,6 +38,7 @@ if (GET('sub') === 'calendar') {
 expose({
   confirmDeleteUser,
   collapse_wrapper,
+  trySetSleepMode,
 })
 
 // Fill the Push Notifications module with life
@@ -93,4 +96,18 @@ async function refreshPushNotificationSettings () {
       throw error
     }
   }, { once: true })
+}
+
+async function trySetSleepMode () {
+  try {
+    const status = parseInt($('#sleep_status').val())
+    const from = $('#sleeprange_from').val()
+    const to = $('#sleeprange_to').val()
+    const message = $('#sleep_msg').val()
+
+    await setSleepStatus(status, from, to, message)
+    pulseSuccess(i18n('settings.sleep.saved'))
+  } catch (e) {
+    pulseError(i18n('error_unexpected'))
+  }
 }