+++ CHANGELOG.md
@@ -69,6 +69,8 @@
 - Rebuild join region to vue !2294 @sefn
 - Rebuild profile menu to vue !2281 @chriswalg
 - Let the community bubble on the map fetch its own content !2311 @alex.simm
+- Changed deleting wall posts to the API #798 !2253 @alex.simm  
+- Use REST API for setting the user's sleep status #798 !2282 @alex.simm
 
 ## Dev/Test/CI stuff
 - bump node to version 16 !2135 @peter.toennies